:art::bug::snake: Convert 'unclassified' taxids of 0 to root taxid of 1
:art::snake: Some taxon-profilers assign unclassified contigs to a taxid of 0 where others assign root taxid of 1.

The NCBI class method of converting the taxid datatype expects a positive integer, so we are converting the unclassified taxid of 0 to
the root taxid of 1. All input predictions will be converted in this same way. the contigs with their converted taxids as well as the number of 'unclassified' taxids
are now emitted in the logs during conversion